production/backend/requirements.txt
flask
flask-cors
plaid-python
pandas
openpyxl
python-dotenv